+++ source/components/FormField.js
@@ -10,39 +10,52 @@ import { composeTheme, addThemeId, didThemePropsChange } from '../utils/themes';
 import { IDENTIFIERS } from '.';
 import type { ThemeContextProp } from './HOC/withTheme';
 
-type Props = {
+export type FormFieldProps = {
   className?: ?string,
   context: ThemeContextProp,
   disabled?: boolean,
   error?: string | Element<any>,
-  inputRef?: ElementRef<*>,
+  formFieldRef: ElementRef<*>,
+  isErrorHidden?: boolean,
+  isErrorShown?: boolean,
+  isShowingErrorOnFocus: boolean,
+  isShowingErrorOnHover: boolean,
   label?: string | Element<any>,
-  render: Function,
+  onChange: Function,
+  render: (setFormFieldRef: (ElementRef<*>) => void) => React$Node,
   skin?: ComponentType<any>,
   theme: ?Object, // will take precedence over theme in context if passed
   themeId: string,
-  themeOverrides: Object
+  themeOverrides: Object,
+  themeVariables?: Object,
 };
 
 type State = {
   error: string,
-  composedTheme: Object
+  composedTheme: Object,
 };
 
-class FormFieldBase extends Component<Props, State> {
+class FormFieldBase extends Component<FormFieldProps, State> {
   // define static properties
   static displayName = 'FormField';
+
   static defaultProps = {
     context: createEmptyContext(),
+    isShowingErrorOnFocus: true,
+    isShowingErrorOnHover: true,
     theme: null,
     themeId: IDENTIFIERS.FORM_FIELD,
-    themeOverrides: {}
+    themeOverrides: {},
   };
 
-  constructor(props: Props) {
+  formFieldRef: ElementRef<*>;
+
+  constructor(props: FormFieldProps) {
     super(props);
 
     const { context, themeId, theme, themeOverrides } = props;
+
+    this.formFieldRef = props.formFieldRef ?? React.createRef();
 
     this.state = {
       error: '',
@@ -50,11 +63,11 @@ class FormFieldBase extends Component<Props, State> {
         addThemeId(theme || context.theme, themeId),
         addThemeId(themeOverrides, themeId),
         context.ROOT_THEME_API
-      )
+      ),
     };
   }
 
-  componentDidUpdate(prevProps: Props) {
+  componentDidUpdate(prevProps: FormFieldProps) {
     if (prevProps !== this.props) {
       didThemePropsChange(prevProps, this.props, this.setState.bind(this));
     }
@@ -63,24 +76,17 @@ class FormFieldBase extends Component<Props, State> {
   setError = (error: string) => this.setState({ error });
 
   focusChild = () => {
-    const { inputRef } = this.props;
-    if (inputRef && inputRef.current) {
-      if (typeof inputRef.current.focus === 'function')
-        inputRef.current.focus();
+    const { formFieldRef } = this;
+    if (formFieldRef && formFieldRef.current) {
+      if (typeof formFieldRef.current.focus === 'function') {
+        formFieldRef.current.focus();
+      }
     }
   };
 
   render() {
     // destructuring props ensures only the "...rest" get passed down
-    const {
-      skin,
-      theme,
-      themeOverrides,
-      error,
-      context,
-      inputRef,
-      ...rest
-    } = this.props;
+    const { skin, theme, themeOverrides, error, context, ...rest } = this.props;
 
     const FormFieldSkin = skin || context.skins[IDENTIFIERS.FORM_FIELD];
 
@@ -89,6 +95,7 @@ class FormFieldBase extends Component<Props, State> {
         error={error || this.state.error}
         setError={this.setError}
         theme={this.state.composedTheme}
+        formFieldRef={this.formFieldRef}
         focusChild={this.focusChild}
         {...rest}
       />