+++ source/skins/simple/OptionsSkin.js
@@ -1,13 +1,14 @@
 // @flow
 import React from 'react';
-import type { Element, ElementRef } from 'react';
+import type { Element, ElementRef, RefObject } from 'react';
 
 // external libraries
 import classnames from 'classnames';
-import { isFunction, isObject } from 'lodash';
+import { isFunction, isObject, escapeRegExp } from 'lodash';
 
 // components
 import { Bubble } from '../../components/Bubble';
+import { Input } from '../../components/Input';
 import { ScrollBar } from '../../components/ScrollBar';
 
 // skins
@@ -17,6 +18,9 @@ type Props = {
   getOptionProps: Function,
   getHighlightedOptionIndex: Function,
   handleClickOnOption: Function,
+  hasSearch?: boolean,
+  hideSearchClearButton?: boolean,
+  highlightSearch?: boolean,
   isHighlightedOption: Function,
   isOpen: boolean,
   isOpeningUpward: boolean,
@@ -25,12 +29,17 @@ type Props = {
   noResults: boolean,
   noResultsMessage: string | Element<any>,
   noSelectedOptionCheckmark?: boolean,
+  onClearSearchValue: Function,
+  onSearch: Function,
   optionHeight: number,
   optionRenderer: Function,
   options: Array<any>,
   optionsRef: ElementRef<*>,
   optionsMaxHeight: number,
   render: Function,
+  searchHeight: number,
+  searchInputRef?: RefObject,
+  searchValue: string,
   selectedOption: any,
   setHighlightedOptionIndex: Function,
   setMouseIsOverOptions?: (boolean) => void,
@@ -44,6 +53,8 @@ export const OptionsSkin = (props: Props) => {
     getOptionProps,
     getHighlightedOptionIndex,
     handleClickOnOption,
+    hasSearch,
+    highlightSearch,
     isHighlightedOption,
     isOpen,
     isOpeningUpward,
@@ -58,6 +69,8 @@ export const OptionsSkin = (props: Props) => {
     options,
     optionsRef,
     render,
+    searchHeight,
+    searchValue,
     setHighlightedOptionIndex,
     setMouseIsOverOptions,
     targetRef,
@@ -66,7 +79,7 @@ export const OptionsSkin = (props: Props) => {
   } = props;
 
   const highlightedOptionIndex = getHighlightedOptionIndex();
-  const isFirstOptionHighlighted = highlightedOptionIndex === 0;
+  const isFirstOptionHighlighted = !hasSearch && highlightedOptionIndex === 0;
   const sortedOptions = isOpeningUpward ? options.slice().reverse() : options;
 
   const renderOptions = () => {
@@ -76,11 +89,15 @@ export const OptionsSkin = (props: Props) => {
     if (!noResults && render) {
       // call user's custom render function
       return render(getOptionProps);
-    } else if (!noResults && !render) {
+    }
+    if (!noResults && !render) {
       // render default simple skin
       return sortedOptions.map((option, index) => {
         // set reference of event handlers in memory to prevent excess re-renders
-        const boundSetHighlightedOptionIndex = setHighlightedOptionIndex.bind(null, index);
+        const boundSetHighlightedOptionIndex = setHighlightedOptionIndex.bind(
+          null,
+          index
+        );
         const boundHandleClickOnOption = handleClickOnOption.bind(null, option);
 
         return (
@@ -91,10 +108,14 @@ export const OptionsSkin = (props: Props) => {
             className={classnames([
               option.className ? option.className : null,
               theme[themeId].option,
-              isHighlightedOption(index) ? theme[themeId].highlightedOption : null,
+              isHighlightedOption(index)
+                ? theme[themeId].highlightedOption
+                : null,
               isSelectedOption(index) ? theme[themeId].selectedOption : null,
               option.isDisabled ? theme[themeId].disabledOption : null,
-              noSelectedOptionCheckmark ? theme[themeId].hasNoSelectedOptionCheckmark : null,
+              noSelectedOptionCheckmark
+                ? theme[themeId].hasNoSelectedOptionCheckmark
+                : null,
             ])}
             onClick={boundHandleClickOnOption}
             onMouseEnter={boundSetHighlightedOptionIndex}
@@ -108,15 +129,51 @@ export const OptionsSkin = (props: Props) => {
     return <li className={theme[themeId].option}>{noResultsMessage}</li>;
   };
 
-  const renderOption = option => {
+  const renderOption = (option) => {
+    const escapedSearchValue = escapeRegExp(searchValue) || '';
     // check if user has passed render prop "optionRenderer"
     if (optionRenderer && isFunction(optionRenderer)) {
       // call user's custom rendering logic
       return optionRenderer(option);
-    } else if (isObject(option)) {
-      return <span className={theme[themeId].label}>{option.label}</span>;
+    }
+    if (isObject(option)) {
+      let { label } = option;
+      // in case `highlightSearch` then `searchValue` is wrapped in an `em` tag
+      if (highlightSearch !== false && escapedSearchValue !== '') {
+        const splitter = new RegExp(`(${escapedSearchValue})`, 'i');
+        const parts = typeof label === 'string' ? label.split(splitter) : label;
+        for (let i = 1; i < parts.length; i += 2) {
+          if (
+            escapeRegExp(parts[i].toLowerCase()) ===
+            `${escapedSearchValue}`.toLowerCase()
+          )
+            parts[i] = <em key={i}>{parts[i]}</em>;
+          label = parts;
+        }
+      }
+      return <span className={theme[themeId].label}>{label}</span>;
     }
     return option;
+  };
+
+  const renderSearch = () => {
+    return (
+      <div className={classnames([theme[themeId].search, 'test'])}>
+        <Input
+          theme={theme}
+          value={searchValue}
+          onChange={props.onSearch}
+          autoFocus
+          inputRef={props.searchInputRef || null}
+        />
+        {!props.hideSearchClearButton && (
+          <button
+            className={searchValue ? theme[themeId].active : null}
+            onClick={props.onClearSearchValue}
+          />
+        )}
+      </div>
+    );
   };
 
   const getScrollBarHeight = (): number => {
@@ -128,9 +185,12 @@ export const OptionsSkin = (props: Props) => {
   };
 
   // Enforce max height of options dropdown if necessary
-  const optionsStyle = optionsMaxHeight == null ? null : {
-    maxHeight: `${optionsMaxHeight}px`
-  };
+  const optionsStyle =
+    optionsMaxHeight == null
+      ? null
+      : {
+          maxHeight: `${optionsMaxHeight}px`,
+        };
 
   return (
     <Bubble
@@ -140,22 +200,28 @@ export const OptionsSkin = (props: Props) => {
         isOpeningUpward ? theme[themeId].openUpward : null,
         isFirstOptionHighlighted && !noResults
           ? theme[themeId].firstOptionHighlighted
-          : null
+          : null,
+        hasSearch ? theme[themeId].hasSearch : null,
       ])}
       isTransparent={false}
       skin={BubbleSkin}
       isOpeningUpward={isOpeningUpward}
       isHidden={!isOpen}
       isFloating
-      noArrow={noOptionsArrow}
+      noArrow={noOptionsArrow || hasSearch}
       targetRef={targetRef}
     >
+      {hasSearch && renderSearch()}
       <ul
         style={optionsStyle}
         ref={optionsRef}
         className={theme[themeId].ul}
-        onMouseEnter={() => setMouseIsOverOptions && setMouseIsOverOptions(true)}
-        onMouseLeave={() => setMouseIsOverOptions && setMouseIsOverOptions(false)}
+        onMouseEnter={() =>
+          setMouseIsOverOptions && setMouseIsOverOptions(true)
+        }
+        onMouseLeave={() =>
+          setMouseIsOverOptions && setMouseIsOverOptions(false)
+        }
       >
         <ScrollBar style={{ height: `${getScrollBarHeight()}px` }}>
           {renderOptions()}